fix: refactor code based on code review.
